requirements.txt
aiogram==2.18
aioschedule==0.5.2
python-dotenv==0.19.2
aiopg==1.3.3
SQLAlchemy==1.4.35
alembic==1.7.7
ayncio==3.4.3